Level manager

+++ SchoolCommon/src/main/java/com/school/common/common/ResourceGet.java
@@ -2,7 +2,7 @@ package com.school.common.common;
 
 public class ResourceGet {
 
-	public String getUserImagePathWithId(Integer id, String photoName) {
+	public String getUserImagePathWithId(Long id, String photoName) {
 		StringBuilder pathImg = new StringBuilder();
 		pathImg.append(SystemConstant.FORWARD_SLASH);
 		pathImg.append(SystemConstant.PHOTOS_OF_USERS_FOLDER);
